Initial refactor into data backend
Updated the API routes to talk to an abstract databackend, rather than directly to the database.

This should make it easier to plug in a different data system (like Palantir)
