+++ api/states.go
@@ -3,7 +3,6 @@ package api
 import (
 	"context"
 	"encoding/json"
-	"fmt"
 	"net/http"
 	"time"
 
@@ -23,18 +22,13 @@ func getStateIDs(w http.ResponseWriter, r *http.Request) {
 
 func getStates(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
-	conn, err := ctx.Value("db").(*pgxpool.Pool).Acquire(ctx)
-	if err != nil {
-		log.Error().Err(err).Msg("Cannot get DB from context")
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-	defer conn.Release()
+	backend := ctx.Value("db").(DataBackend)
 
-	query := "SELECT c.statefp, c.State, s.Update, SUM(s.Confirmed) as confirmed, SUM(s.NewConfirmed), SUM(s.Dead), SUM(s.NewDead) FROM counties as c " +
-		"LEFT JOIN cases as s " +
-		"ON s.geoid = c.ID "
+	// query := "SELECT c.statefp, c.State, s.Update, SUM(s.Confirmed) as confirmed, SUM(s.NewConfirmed), SUM(s.Dead), SUM(s.NewDead) FROM counties as c " +
+	// 	"LEFT JOIN cases as s " +
+	// 	"ON s.geoid = c.ID "
 
+	var t *time.Time = nil
 	start, ok := r.URL.Query()["start"]
 	if ok && len(start) == 1 {
 		startTime, err := time.Parse(time.RFC3339, start[0])
@@ -43,12 +37,20 @@ func getStates(w http.ResponseWriter, r *http.Request) {
 			w.WriteHeader(http.StatusInternalServerError)
 			return
 		}
-		query = fmt.Sprintf("%s WHERE s.update > '%s' ", query, startTime.Format(time.RFC3339))
+		t = &startTime
 	}
-	query = query + "GROUP BY c.statefp, c.state, s.update " +
-		"ORDER BY c.state, s.update;"
+	// 	query = fmt.Sprintf("%s WHERE s.update > '%s' ", query, startTime.Format(time.RFC3339))
+	// }
+	// query = query + "GROUP BY c.statefp, c.state, s.update " +
+	// 	"ORDER BY c.state, s.update;"
 
-	cases, err := queryStateCases(ctx, conn, query)
+	// cases, err := queryStateCases(ctx, conn, query)
+	// if err != nil {
+	// 	log.Error().Err(err).Msg("Cannot query for state cases")
+	// 	w.WriteHeader(http.StatusInternalServerError)
+	// 	return
+	// }
+	cases, err := backend.GetTopStates(ctx, t)
 	if err != nil {
 		log.Error().Err(err).Msg("Cannot query for state cases")
 		w.WriteHeader(http.StatusInternalServerError)
@@ -94,23 +96,30 @@ func getStateGeo(w http.ResponseWriter, r *http.Request) {
 func getStateCases(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	stateID := ctx.Value("stateID").(string)
-	conn, err := ctx.Value("db").(*pgxpool.Pool).Acquire(ctx)
-	if err != nil {
-		log.Error().Err(err).Msg("Cannot get DB from context")
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-	defer conn.Release()
+	backend := ctx.Value("db").(DataBackend)
+	// conn, err := ctx.Value("db").(*pgxpool.Pool).Acquire(ctx)
+	// if err != nil {
+	// 	log.Error().Err(err).Msg("Cannot get DB from context")
+	// 	w.WriteHeader(http.StatusInternalServerError)
+	// 	return
+	// }
+	// defer conn.Release()
 
-	log.Debug().Msgf("Returning case data for state: %s\n", stateID)
+	// log.Debug().Msgf("Returning case data for state: %s\n", stateID)
 
-	query := "SELECT c.statefp, c.state, a.update, SUM(a.confirmed) as confirmed, SUM(a.newconfirmed), SUM(a.dead), SUM(a.newdead) from counties as c " +
-		"LEFT JOIN cases as a ON c.id = a.geoid " +
-		"WHERE c.statefp = $1 " +
-		"GROUP BY c.statefp, c.state, a.update " +
-		"ORDER BY a.update DESC, confirmed DESC"
+	// query := "SELECT c.statefp, c.state, a.update, SUM(a.confirmed) as confirmed, SUM(a.newconfirmed), SUM(a.dead), SUM(a.newdead) from counties as c " +
+	// 	"LEFT JOIN cases as a ON c.id = a.geoid " +
+	// 	"WHERE c.statefp = $1 " +
+	// 	"GROUP BY c.statefp, c.state, a.update " +
+	// 	"ORDER BY a.update DESC, confirmed DESC"
 
-	cases, err := queryStateCases(ctx, conn, query, stateID)
+	// cases, err := queryStateCases(ctx, conn, query, stateID)
+	// if err != nil {
+	// 	log.Error().Err(err).Msg("Cannot query state cases")
+	// 	w.WriteHeader(http.StatusInternalServerError)
+	// 	return
+	// }
+	cases, err := backend.GetStateCases(ctx, stateID)
 	if err != nil {
 		log.Error().Err(err).Msg("Cannot query state cases")
 		w.WriteHeader(http.StatusInternalServerError)
@@ -125,42 +134,42 @@ func getStateCases(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-func queryStateCases(ctx context.Context, conn *pgxpool.Conn, sql string, args ...interface{}) ([]cmd.StateCases, error) {
-	cases := make([]cmd.StateCases, 0)
+// func queryStateCases(ctx context.Context, conn *pgxpool.Conn, sql string, args ...interface{}) ([]cmd.StateCases, error) {
+// 	cases := make([]cmd.StateCases, 0)
 
-	rows, err := conn.Query(ctx, sql, args...)
-	if err != nil {
-		return cases, err
-	}
+// 	rows, err := conn.Query(ctx, sql, args...)
+// 	if err != nil {
+// 		return cases, err
+// 	}
 
-	for rows.Next() {
-		var id string
-		var state string
-		var confirmed int
-		var newConfirmed int
-		var dead int
-		var newDead int
-		var reported time.Time
-		err := rows.Scan(&id, &state, &reported, &confirmed, &newConfirmed, &dead, &newDead)
-		if err != nil {
-			return cases, err
-		}
-		caseCount := &cmd.CaseCount{
-			Confirmed:    confirmed,
-			NewConfirmed: newConfirmed,
-			Dead:         dead,
-			NewDead:      newDead,
-			Reported:     reported,
-		}
+// 	for rows.Next() {
+// 		var id string
+// 		var state string
+// 		var confirmed int
+// 		var newConfirmed int
+// 		var dead int
+// 		var newDead int
+// 		var reported time.Time
+// 		err := rows.Scan(&id, &state, &reported, &confirmed, &newConfirmed, &dead, &newDead)
+// 		if err != nil {
+// 			return cases, err
+// 		}
+// 		caseCount := &cmd.CaseCount{
+// 			Confirmed:    confirmed,
+// 			NewConfirmed: newConfirmed,
+// 			Dead:         dead,
+// 			NewDead:      newDead,
+// 			Reported:     reported,
+// 		}
 
-		cases = append(cases, cmd.StateCases{
-			ID:        id,
-			State:     state,
-			CaseCount: caseCount,
-		})
-	}
-	return cases, nil
-}
+// 		cases = append(cases, cmd.StateCases{
+// 			ID:        id,
+// 			State:     state,
+// 			CaseCount: caseCount,
+// 		})
+// 	}
+// 	return cases, nil
+// }
 
 func stateCTX(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
