+++ go.mod
@@ -9,6 +9,5 @@ require (
 	github.com/jackc/pgx/v4 v4.5.0
 	github.com/rs/zerolog v1.15.0
 	github.com/stretchr/testify v1.5.1
-	github.com/rs/zerolog v1.15.0
 	github.com/urfave/cli/v2 v2.2.0
 )
